+++ tests/test.rs
@@ -8,28 +8,28 @@ macro_rules! help_msg {
             "Usage:\n",
             "  aki-mcolor [options]\n",
             "\n",
-            "color marker by rust lang.\n",
+            "mark up text with color\n",
             "\n",
             "Options:\n",
-            "  -r, --red <exp>      regular expression, mark color is red\n",
-            "  -g, --green <exp>    regular expression, mark color is green\n",
-            "  -b, --blue <exp>     regular expression, mark color is blue\n",
-            "  -c, --cyan <exp>     regular expression, mark color is cyan\n",
-            "  -m, --magenda <exp>  regular expression, mark color is magenda\n",
-            "  -y, --yellow <exp>   regular expression, mark color is yellow\n",
-            "  -u, --unmark <exp>   regular expression, unmark color\n",
+            "  -r, --red <exp>       regular expression, mark color is red\n",
+            "  -g, --green <exp>     regular expression, mark color is green\n",
+            "  -b, --blue <exp>      regular expression, mark color is blue\n",
+            "  -c, --cyan <exp>      regular expression, mark color is cyan\n",
+            "  -m, --magenda <exp>   regular expression, mark color is magenda\n",
+            "  -y, --yellow <exp>    regular expression, mark color is yellow\n",
+            "  -u, --unmark <exp>    regular expression, unmark color\n",
             "\n",
-            "  -H, --help     display this help and exit\n",
-            "  -V, --version  display version information and exit\n",
+            "  -H, --help        display this help and exit\n",
+            "  -V, --version     display version information and exit\n",
             "\n",
             "Env:\n",
-            "  RUST_MCOLOR_RED_ST     red start sequence\n",
-            "  RUST_MCOLOR_GREEN_ST   greep start sequence\n",
-            "  RUST_MCOLOR_BLUE_ST    blue start sequence\n",
-            "  RUST_MCOLOR_CYAN_ST    cyan start sequence\n",
-            "  RUST_MCOLOR_MAGENDA_ST magenda start sequence\n",
-            "  RUST_MCOLOR_YELLOW_ST  yellow start sequence\n",
-            "  RUST_MCOLOR_ED         color end sequence\n",
+            "  AKI_MCOLOR_RED_ST         red start sequence\n",
+            "  AKI_MCOLOR_GREEN_ST       greep start sequence\n",
+            "  AKI_MCOLOR_BLUE_ST        blue start sequence\n",
+            "  AKI_MCOLOR_CYAN_ST        cyan start sequence\n",
+            "  AKI_MCOLOR_MAGENDA_ST     magenda start sequence\n",
+            "  AKI_MCOLOR_YELLOW_ST      yellow start sequence\n",
+            "  AKI_MCOLOR_ED             color end sequence\n",
             "\n"
         )
     };
@@ -121,8 +121,8 @@ mod test_1 {
     #[test]
     fn test_red() {
         let mut env: HashMap<String, String> = HashMap::new();
-        env.insert("RUST_MCOLOR_RED_ST".to_string(), "<S>".to_string());
-        env.insert("RUST_MCOLOR_ED".to_string(), "<E>".to_string());
+        env.insert("AKI_MCOLOR_RED_ST".to_string(), "<S>".to_string());
+        env.insert("AKI_MCOLOR_ED".to_string(), "<E>".to_string());
         let oup = exec_target_with_env_in(TARGET_EXE_PATH, &["-r", "c"], env, b"abcdefg" as &[u8]);
         assert_eq!(oup.stderr, "");
         assert_eq!(oup.stdout, "ab<S>c<E>defg\n");
@@ -131,8 +131,8 @@ mod test_1 {
     #[test]
     fn test_green() {
         let mut env: HashMap<String, String> = HashMap::new();
-        env.insert("RUST_MCOLOR_GREEN_ST".to_string(), "<S>".to_string());
-        env.insert("RUST_MCOLOR_ED".to_string(), "<E>".to_string());
+        env.insert("AKI_MCOLOR_GREEN_ST".to_string(), "<S>".to_string());
+        env.insert("AKI_MCOLOR_ED".to_string(), "<E>".to_string());
         let oup = exec_target_with_env_in(TARGET_EXE_PATH, &["-g", "c"], env, b"abcdefg" as &[u8]);
         assert_eq!(oup.stderr, "");
         assert_eq!(oup.stdout, "ab<S>c<E>defg\n");
@@ -141,8 +141,8 @@ mod test_1 {
     #[test]
     fn test_blue() {
         let mut env: HashMap<String, String> = HashMap::new();
-        env.insert("RUST_MCOLOR_BLUE_ST".to_string(), "<S>".to_string());
-        env.insert("RUST_MCOLOR_ED".to_string(), "<E>".to_string());
+        env.insert("AKI_MCOLOR_BLUE_ST".to_string(), "<S>".to_string());
+        env.insert("AKI_MCOLOR_ED".to_string(), "<E>".to_string());
         let oup = exec_target_with_env_in(TARGET_EXE_PATH, &["-b", "c"], env, b"abcdefg" as &[u8]);
         assert_eq!(oup.stderr, "");
         assert_eq!(oup.stdout, "ab<S>c<E>defg\n");
@@ -151,8 +151,8 @@ mod test_1 {
     #[test]
     fn test_cyan() {
         let mut env: HashMap<String, String> = HashMap::new();
-        env.insert("RUST_MCOLOR_CYAN_ST".to_string(), "<S>".to_string());
-        env.insert("RUST_MCOLOR_ED".to_string(), "<E>".to_string());
+        env.insert("AKI_MCOLOR_CYAN_ST".to_string(), "<S>".to_string());
+        env.insert("AKI_MCOLOR_ED".to_string(), "<E>".to_string());
         let oup = exec_target_with_env_in(TARGET_EXE_PATH, &["-c", "c"], env, b"abcdefg" as &[u8]);
         assert_eq!(oup.stderr, "");
         assert_eq!(oup.stdout, "ab<S>c<E>defg\n");
@@ -161,8 +161,8 @@ mod test_1 {
     #[test]
     fn test_magenda() {
         let mut env: HashMap<String, String> = HashMap::new();
-        env.insert("RUST_MCOLOR_MAGENDA_ST".to_string(), "<S>".to_string());
-        env.insert("RUST_MCOLOR_ED".to_string(), "<E>".to_string());
+        env.insert("AKI_MCOLOR_MAGENDA_ST".to_string(), "<S>".to_string());
+        env.insert("AKI_MCOLOR_ED".to_string(), "<E>".to_string());
         let oup = exec_target_with_env_in(TARGET_EXE_PATH, &["-m", "c"], env, b"abcdefg" as &[u8]);
         assert_eq!(oup.stderr, "");
         assert_eq!(oup.stdout, "ab<S>c<E>defg\n");
@@ -171,8 +171,8 @@ mod test_1 {
     #[test]
     fn test_yellow() {
         let mut env: HashMap<String, String> = HashMap::new();
-        env.insert("RUST_MCOLOR_YELLOW_ST".to_string(), "<S>".to_string());
-        env.insert("RUST_MCOLOR_ED".to_string(), "<E>".to_string());
+        env.insert("AKI_MCOLOR_YELLOW_ST".to_string(), "<S>".to_string());
+        env.insert("AKI_MCOLOR_ED".to_string(), "<E>".to_string());
         let oup = exec_target_with_env_in(TARGET_EXE_PATH, &["-y", "c"], env, b"abcdefg" as &[u8]);
         assert_eq!(oup.stderr, "");
         assert_eq!(oup.stdout, "ab<S>c<E>defg\n");