+++ CHANGELOG.md
@@ -2,6 +2,14 @@ TBD
 ===
 Unreleased changes. Release notes have not yet been written.
 
+0.1.15 (2021-03-02)
+=====
+Feature:
+
+* change env: RUST_MCOLOR_RED_ST to AKI_MCOLOR_RED_ST
+* update crate: flood-tide-gen
+* add some documents
+
 0.1.14 (2021-02-22)
 =====
 Feature: